+++ src/WebHemi/Middleware/Action/Admin/ControlPanel/Themes/ViewAction.php
@@ -1,0 +1,43 @@
+<?php
+/**
+ * WebHemi.
+ *
+ * PHP version 7.1
+ *
+ * @copyright 2012 - 2017 Gixx-web (http://www.gixx-web.com)
+ * @license   https://opensource.org/licenses/MIT The MIT License (MIT)
+ *
+ * @link      http://www.gixx-web.com
+ */
+declare(strict_types = 1);
+
+namespace WebHemi\Middleware\Action\Admin\ControlPanel\Themes;
+
+use RuntimeException;
+use WebHemi\Middleware\Action\AbstractMiddlewareAction;
+
+/**
+ * Class ViewAction
+ */
+class ViewAction extends AbstractMiddlewareAction
+{
+    /**
+     * Gets template map name or template file path.
+     *
+     * @return string
+     */
+    public function getTemplateName() : string
+    {
+        return 'admin-control-panel-themes-view';
+    }
+
+    /**
+     * Gets template data.
+     *
+     * @return array
+     */
+    public function getTemplateData() : array
+    {
+        return [];
+    }
+}